Refactor Ambrogio class and related files
This commit refactors the Ambrogio class and related files. It introduces new imports for the FeatureExtractor and DataSetManager classes. It also adds a new method train() for training the model and a backpropagation() method for updating the weights of the connections. Additionally, it includes a new method getConnectionsOfNeuron() for retrieving the connections of a neuron. The commit also updates the output() method in the Neuron class to use the sigmoid activation function. Finally, it updates the calcCrossEntropyLoss() method in the SoftMaxNeuron class to avoid log(0) errors.